V2 of the code
I’ve modified the code to add the following:

    Used the OneButton library so that the setpoint can be increased by
clicking button and decreased by long clicking the button
    Used the Timer library so that I don’t rely on delay – which was
inaccurate
    Change the logic for temperature such that the heater gets switched
on if the temperature goes below the setpoint. Makes more sense as you
don’t need the heater to be on at the setpoint. Also, reduces chances
of the heater malfunctioning by switching on and off all the time
    Used a similar logic for humidity – on when humidity is less than
50 and off when it is more than 60